requirements.txt
yfinance==0.2.44
pandas==2.2.2
numpy==2.1.0
matplotlib==3.9.2
seaborn==0.13.2
scikit-learn==1.5.2
ta==0.11.0
requests==2.32.3
beautifulsoup4==4.12.3
textblob==0.17.1
vaderSentiment==3.3.2
streamlit==1.39.0
plotly==5.24.1
openai==1.51.0
python-dotenv==1.0.1
schedule==1.2.2
flask==3.0.3
flask-cors==4.0.1
deepseek==1.0.0
akshare==1.12.85
gunicorn==21.2.0
psutil==5.9.8
